Extracted data structures in TheoryObject
